Fix PIR and change serial prints

+++ firmware/sensor/src/sensor.cpp
@@ -332,9 +332,9 @@ void evaluateState() {
 
 void readRemoteState() {
     if (HWSERIAL.available() > 0) {
-        int incomingByte;
+        uint8_t incomingByte;
         incomingByte = HWSERIAL.read();
-        // Serial.println(incomingByte);
+        // Serial.println(incomingByte, HEX);
         
         if (incomingByte == 'O') {
             remoteState = STATE_OPEN;
